refactor servers_list
