+++ erddapy/__init__.py
@@ -1,11 +1,9 @@
 """Easier access to scientific data."""
 
 from erddapy.erddapy import ERDDAP
-from erddapy.servers.servers import servers
 
 __all__ = [
     "ERDDAP",
-    "servers",
 ]
 
 try:
